Extracts 'getExistingDocSchema' to 'SchemaImporter.ts'
